+++ src/main/java/com/kruemblegard/client/render/PebblitRenderer.java
@@ -1,23 +1,16 @@
 package com.kruemblegard.client.render;
 
-import com.kruemblegard.Kruemblegard;
+import com.kruemblegard.client.render.model.PebblitModel;
+import com.kruemblegard.entity.PebblitEntity;
 
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
-import net.minecraft.client.renderer.entity.SilverfishRenderer;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.monster.Silverfish;
 
-public class PebblitRenderer extends SilverfishRenderer {
+import software.bernie.geckolib.renderer.GeoEntityRenderer;
 
-    private static final ResourceLocation PEBBLIT_TEXTURE =
-            new ResourceLocation(Kruemblegard.MOD_ID, "textures/entity/pebblit.png");
+public class PebblitRenderer extends GeoEntityRenderer<PebblitEntity> {
 
     public PebblitRenderer(EntityRendererProvider.Context context) {
-        super(context);
-    }
-
-    @Override
-    public ResourceLocation getTextureLocation(Silverfish entity) {
-        return PEBBLIT_TEXTURE;
+        super(context, new PebblitModel());
+        this.shadowRadius = 0.35f;
     }
 }